+++ packages/test-common/CHANGELOG.md
@@ -83,6 +83,25 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [0.5.36](https://github.com/rarible/flow-sdk/compare/v0.5.35...v0.5.36) (2022-02-18)
+
+**Note:** Version bump only for package @rarible/flow-test-common
+
+
+
+
+
+## [0.5.34](https://github.com/rarible/flow-sdk/compare/v0.5.33...v0.5.34) (2022-02-16)
+
+
+### Bug Fixes
+
+* replace get royalties method ([#65](https://github.com/rarible/flow-sdk/issues/65)) ([6db4d54](https://github.com/rarible/flow-sdk/commit/6db4d5417f441e50091d3914d390daf5f89c1022))
+
+
+
+
+
 ## [0.5.31](https://github.com/rarible/flow-sdk/compare/v0.5.30...v0.5.31) (2022-01-31)